+++ lib/alt_screens/settings/about.dart
@@ -14,7 +14,7 @@ Widget _showAbout() {
   return AboutPage(
     applicationName: "AssignGo",
     values: {
-      'version': 'Alpha 1.2',
+      'version': '0.2.0',
       'year': DateTime.now().year.toString(),
     },
     applicationLegalese: 'Copyright © Ishan Misra, {{ year }}',